+++ app/src/main/AndroidManifest.xml
@@ -13,10 +13,10 @@
         android:theme="@style/Theme.Movie"
         tools:targetApi="31">
         <activity
-            android:name=".view.ReservationActivity"
+            android:name=".view.reservation.ReservationActivity"
             android:exported="false" />
         <activity
-            android:name=".view.TicketActivity"
+            android:name=".view.ticket.TicketActivity"
             android:exported="false" />
         <activity
             android:name=".view.MainActivity"
