+++ app/src/main/java/woowacourse/movie/domain/screening/ShowtimePolicy.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.domain
+package woowacourse.movie.domain.screening
 
 import java.time.DayOfWeek
 import java.time.LocalDate
