+++ app/src/main/java/woowacourse/movie/view/reservation/ShowReservationDialog.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.view
+package woowacourse.movie.view.reservation
 
 import android.content.Context
 import android.content.DialogInterface
