+++ app/src/main/java/woowacourse/movie/view/screening/adapter/ScreeningAdapter.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.view
+package woowacourse.movie.view.screening.adapter
 
 import android.view.LayoutInflater
 import android.view.View
@@ -8,9 +8,9 @@ import android.widget.Button
 import android.widget.ImageView
 import android.widget.TextView
 import woowacourse.movie.R
-import woowacourse.movie.view.model.Screening
+import woowacourse.movie.view.reservation.model.Screening
 
-class MovieAdapter(
+class ScreeningAdapter(
     screenings: List<Screening>,
     private val onClickReserveButton: (
         Screening,
@@ -32,7 +32,7 @@ class MovieAdapter(
         val view =
             convertView ?: LayoutInflater
                 .from(parent?.context)
-                .inflate(R.layout.item_movie, parent, false)
+                .inflate(R.layout.item_screening, parent, false)
         val screening: Screening = screenings[position]
         initMovieItemView(view, screening)
 
@@ -44,10 +44,10 @@ class MovieAdapter(
         screening: Screening,
     ) {
         with(screening) {
-            val titleView = view.findViewById<TextView>(R.id.tv_item_movie_title)
+            val titleView = view.findViewById<TextView>(R.id.tv_item_screening_title)
             titleView.text = screening.title
 
-            val screeningDateView = view.findViewById<TextView>(R.id.tv_item_movie_screening_date)
+            val screeningDateView = view.findViewById<TextView>(R.id.tv_item_screening_date)
             screeningDateView.text =
                 view.context.getString(
                     R.string.screening_period,
@@ -59,14 +59,14 @@ class MovieAdapter(
                     endDay,
                 )
 
-            val runningTimeView = view.findViewById<TextView>(R.id.tv_item_movie_running_time)
+            val runningTimeView = view.findViewById<TextView>(R.id.tv_item_screening_running_time)
             runningTimeView.text =
                 view.context.getString(R.string.running_time, runningTime)
 
-            val posterView = view.findViewById<ImageView>(R.id.iv_item_movie_poster)
+            val posterView = view.findViewById<ImageView>(R.id.iv_item_screening_poster)
             posterView.setImageResource(posterId)
 
-            val reserveButton = view.findViewById<Button>(R.id.btn_item_movie_reserve)
+            val reserveButton = view.findViewById<Button>(R.id.btn_item_screening_reserve)
             reserveButton.setOnClickListener {
                 onClickReserveButton(
                     screening,
