+++ app/src/main/res/layout/activity_main.xml
@@ -5,4 +5,4 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     tools:context=".view.MainActivity"
-    tools:listitem="@layout/item_movie" />
+    tools:listitem="@layout/item_screening" />
