+++ app/src/main/res/layout/item_screening.xml
@@ -8,7 +8,7 @@
     android:padding="12dp">
 
     <ImageView
-        android:id="@+id/iv_item_movie_poster"
+        android:id="@+id/iv_item_screening_poster"
         android:layout_width="76dp"
         android:layout_height="wrap_content"
         android:layout_marginEnd="15dp"
@@ -22,7 +22,7 @@
         android:orientation="vertical">
 
         <TextView
-            android:id="@+id/tv_item_movie_title"
+            android:id="@+id/tv_item_screening_title"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginBottom="13dp"
@@ -32,7 +32,7 @@
             tools:text="해리 포터와 마법사의 돌" />
 
         <TextView
-            android:id="@+id/tv_item_movie_screening_date"
+            android:id="@+id/tv_item_screening_date"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginBottom="4dp"
@@ -41,7 +41,7 @@
             tools:text="@string/screening_period" />
 
         <TextView
-            android:id="@+id/tv_item_movie_running_time"
+            android:id="@+id/tv_item_screening_running_time"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginBottom="6dp"
@@ -50,13 +50,13 @@
             tools:text="@string/running_time" />
 
         <Button
-            android:id="@+id/btn_item_movie_reserve"
+            android:id="@+id/btn_item_screening_reserve"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_gravity="end"
             android:paddingHorizontal="16dp"
             android:paddingVertical="10dp"
-            android:text="@string/item_movie_reserve"
+            android:text="@string/item_screening_reserve"
             android:textColor="@color/white"
             android:textSize="14sp"
             android:textStyle="bold" />
